introduced and implemented new Router

+++ src/Controller/DefaultController.php
@@ -13,7 +13,8 @@ class DefaultController extends AbstractController
         $this->display('index.html', ['results' => $results]);
     }
 
-    public function postShow($id) {
+    public function postShow() {
+        $id = $this->query['post_id'];
         $uid = $this->getUserId();
 
         # retrieve results
@@ -31,6 +32,9 @@ class DefaultController extends AbstractController
             ;
             ")->fetch(\PDO::FETCH_ASSOC);
 
+        if (!$result) {
+            throw new \Exception("This Resource does not exist", 404);
+        }
         if ($uid) {
             $result2 = $this->getDbConnection()->query("
             SELECT `id` FROM `post_likes` WHERE `post`='$id' AND `user`='$uid';
@@ -41,19 +45,23 @@ class DefaultController extends AbstractController
         $this->display('post.html', ['result' => $result]);
     }
 
-    public function postEdit($id) {
-        # retrieve results
-        $result = $this->getDbConnection()->query("SELECT * FROM `posts` WHERE `id`='$id';")->fetch(\PDO::FETCH_ASSOC);
-        if ($result) {
-            if ($result['user'] !== $this->getUserId()) {
-                throw new \Exception('Users are not allowed to edit foreign Posts', 403);
+    public function postEdit() {
+        $id = $this->query['post_id'] ?? 0;
+        if ($id) {
+            # retrieve results
+            $result = $this->getDbConnection()->query("SELECT * FROM `posts` WHERE `id`='$id';")->fetch(\PDO::FETCH_ASSOC);
+            if ($result) {
+                if ($result['user'] !== $this->getUserId()) {
+                    throw new \Exception('Users are not allowed to edit foreign Posts', 403);
+                }
+            } else {
+                throw new \Exception('This Post does not exist', 404);
             }
-        } else if ($id) {
-            throw new \Exception('This Post does not exist', 404);
         }
-        $this->display('post_update.html', ['result' => $result]);        
+        $this->display('post_update.html', ['result' => $result ?? []]);        
     }
-    public function postSave($data) {
+    public function postSave() {
+        $data = $_REQUEST;
         # TODO: perform mysql query sanitation
         $user = $this->getUser();
         if (!$user) {
@@ -94,10 +102,8 @@ class DefaultController extends AbstractController
             return $this->getDbConnection()->exec($sql);
         }
     }
-    public function postKill($id) {
-        if (!$id || !is_numeric($id)) {
-            return;
-        }
+    public function postKill() {
+        $id = $this->query['post_id'];
         $uid = $this->getUserId();
         if (!$uid) {
             throw new \Exception("This resource is only available for authenticated users", 401);
@@ -106,18 +112,36 @@ class DefaultController extends AbstractController
         return $this->index();
     }
 
-    public function postLikeSave($postId) {
+    public function postLikeSave() {
+        $postId = $this->query['post_id'];
         $uid = $this->getUserId();
-        $result = $this->getDbConnection()->exec("INSERT INTO `post_likes` 
-        (`post`,`user`,`created_at`) VALUES 
-        ('$postId', '$uid', '".date('Y-m-d H:i:s')."');
-        ");
+        if (!$uid) {
+            throw new \Exception("This resource is only available for authenticated users", 401);
+        }
+
+        try {
+            $result = $this->getDbConnection()->exec("INSERT INTO `post_likes` 
+            (`post`,`user`,`created_at`) VALUES 
+            ('$postId', '$uid', '".date('Y-m-d H:i:s')."');
+            ");            
+        } catch (\PDOException $e) {
+            throw new \Exception($e->getMessage(),304);
+        }
         return $result;
     }
 
-    public function postLikeDelete($postId) {
+    public function postLikeDelete() {
+        $postId = $this->query['post_id'];
         $uid = $this->getUserId();
-        $result = $this->getDbConnection()->exec("DELETE FROM `post_likes` WHERE `post`=$postId AND `user`=$uid;");
+        if (!$uid) {
+            throw new \Exception("This resource is only available for authenticated users", 401);
+        }
+
+        try {
+            $result = $this->getDbConnection()->exec("DELETE FROM `post_likes` WHERE `post`=$postId AND `user`=$uid;");
+        } catch (\PDOException $e) {
+            throw new \Exception($e->getMessage(),304);
+        }
         return $result;
     }
